LayoutTests/http/tests/inspector/network/network-disable-cache-memory.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
function loadImage(callback)
{
    var image = new Image();
    image.onload = callback;
    image.src = "resources/resource.php?type=image&random=1&cached=1";
    document.body.appendChild(image);
}

function loadData()
{
    loadImage(imageLoaded1);
}

function imageLoaded1()
{
    loadImage(imageLoaded2);
}

function imageLoaded2()
{
    console.log("First two images loaded.");
}

function cacheDisabled()
{
    loadImage(allImagesLoaded);
}

function allImagesLoaded()
{
    console.log("Done.");
}

function test()
{
    InspectorTest.addSniffer(WebInspector.ConsoleView.prototype, "addMessage", step2);
    InspectorTest.evaluateInPage("loadData()");

    function step2(msg)
    {
        NetworkAgent.setCacheDisabled(true, step3);
    }

    function step3(msg)
    {
        InspectorTest.addSniffer(WebInspector.ConsoleView.prototype, "addMessage", step4);
        InspectorTest.evaluateInPage("cacheDisabled()");
    }

    function step4(msg)
    {
        // inspector-test.js appears in network panel occasionally in Safari on
        // Mac, so checking two last resources.
        var resourcesCount = WebInspector.panels.network.resources.length;
        var resource1 = WebInspector.panels.network.resources[resourcesCount - 2];
        var resource2 = WebInspector.panels.network.resources[resourcesCount - 1];
        resource1.requestContent(contentLoaded);
        resource2.requestContent(contentLoaded);
    }

    var contentLoadedCount = 0;
    function contentLoaded()
    {
        if (++contentLoadedCount !== 2)
            return;

        var resourcesCount = WebInspector.panels.network.resources.length;
        var resource1 = WebInspector.panels.network.resources[resourcesCount - 2];
        var resource2 = WebInspector.panels.network.resources[resourcesCount - 1];

        InspectorTest.addResult(resource1.url);
        InspectorTest.addResult(resource2.url);
        InspectorTest.assertTrue(resource1.content !== resource2.content, "Second and third resources differ");
        NetworkAgent.setCacheDisabled(false, step5);
    }

    function step5(msg)
    {
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
    <p>Tests disabling cache from inspector.</p>
</body>
</html>
